requestclient: explicitly compare fields instead
