+++ internal/requestclient/grpc_test.go
@@ -1,40 +1,38 @@
-package requestclient_test
+package requestclient
 
 import (
 	"context"
 	"net"
 	"testing"
 
-	"github.com/google/go-cmp/cmp"
-	"github.com/google/go-cmp/cmp/cmpopts"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 	"google.golang.org/grpc/peer"
-
-	"github.com/sourcegraph/sourcegraph/internal/requestclient"
 )
 
 func TestPropagator(t *testing.T) {
 	tests := []struct {
 		name string
 
-		requestClient *requestclient.Client
+		requestClient *Client
 		requestPeer   *peer.Peer
 
-		wantClient *requestclient.Client
+		wantClient *Client
 	}{
 		{
 			name: "no client or peer",
 
-			wantClient: &requestclient.Client{},
+			wantClient: &Client{},
 		},
 
 		{
 			name: "client with no peer",
-			requestClient: &requestclient.Client{
+			requestClient: &Client{
 				IP:           "192.168.1.1",
 				ForwardedFor: "192.168.1.2",
 			},
 
-			wantClient: &requestclient.Client{
+			wantClient: &Client{
 				IP:           "192.168.1.1",
 				ForwardedFor: "192.168.1.2",
 			},
@@ -46,19 +44,19 @@ func TestPropagator(t *testing.T) {
 				Addr: &net.IPAddr{IP: net.ParseIP("192.168.1.1")},
 			},
 
-			wantClient: &requestclient.Client{
+			wantClient: &Client{
 				IP: "192.168.1.1",
 			},
 		},
 		{
 			name: "peer only (non-nil empty client)",
 
-			requestClient: &requestclient.Client{},
+			requestClient: &Client{},
 			requestPeer: &peer.Peer{
 				Addr: &net.IPAddr{IP: net.ParseIP("192.168.1.1")},
 			},
 
-			wantClient: &requestclient.Client{
+			wantClient: &Client{
 				IP: "192.168.1.1",
 			},
 		},
@@ -66,7 +64,7 @@ func TestPropagator(t *testing.T) {
 		{
 			name: "client should override peer",
 
-			requestClient: &requestclient.Client{
+			requestClient: &Client{
 				IP:           "192.168.1.1",
 				ForwardedFor: "192.168.1.2",
 			},
@@ -74,7 +72,7 @@ func TestPropagator(t *testing.T) {
 				Addr: &net.IPAddr{IP: net.ParseIP("192.168.1.3")},
 			},
 
-			wantClient: &requestclient.Client{
+			wantClient: &Client{
 				IP:           "192.168.1.1",
 				ForwardedFor: "192.168.1.2",
 			},
@@ -83,14 +81,14 @@ func TestPropagator(t *testing.T) {
 		{
 			name: "client for ForwardedFor, peer for IP",
 
-			requestClient: &requestclient.Client{
+			requestClient: &Client{
 				ForwardedFor: "192.168.1.2",
 			},
 			requestPeer: &peer.Peer{
 				Addr: &net.IPAddr{IP: net.ParseIP("192.168.1.3")},
 			},
 
-			wantClient: &requestclient.Client{
+			wantClient: &Client{
 				IP:           "192.168.1.3",
 				ForwardedFor: "192.168.1.2",
 			},
@@ -101,26 +99,64 @@ func TestPropagator(t *testing.T) {
 		t.Run(test.name, func(t *testing.T) {
 			requestCtx := context.Background()
 			if test.requestClient != nil {
-				requestCtx = requestclient.WithClient(requestCtx, test.requestClient)
+				requestCtx = WithClient(requestCtx, test.requestClient)
 			}
 
 			if test.requestPeer != nil {
 				requestCtx = peer.NewContext(requestCtx, test.requestPeer)
 			}
 
-			propagator := &requestclient.Propagator{}
+			propagator := &Propagator{}
 			md := propagator.FromContext(requestCtx)
 
 			resultCtx := propagator.InjectContext(requestCtx, md)
-			if diff := cmp.Diff(
-				test.wantClient,
-				requestclient.FromContext(resultCtx),
-				// Ignore unexported fields in Client - this test only tests
-				// the exported API surface (hence why we run this test in
-				// package requestclient_test)
-				cmpopts.IgnoreUnexported(requestclient.Client{}),
-			); diff != "" {
-				t.Errorf("Client mismatch (-want +got):\n%s", diff)
+
+			// Explicitly compare exported fields because cmp.Diff doesn't work
+			// when there are unexported fields
+			rc := FromContext(resultCtx)
+			require.NotNil(t, rc)
+			assert.Equal(t, test.wantClient.IP, rc.IP)
+			assert.Equal(t, test.wantClient.ForwardedFor, rc.ForwardedFor)
+			assert.Equal(t, test.wantClient.UserAgent, rc.UserAgent)
+		})
+	}
+}
+
+func TestBaseIP(t *testing.T) {
+	tests := []struct {
+		name string
+		addr net.Addr
+		want string
+	}{
+		{
+			name: "TCP address",
+			addr: &net.TCPAddr{
+				IP:   net.ParseIP("127.0.127.2"),
+				Port: 448,
+			},
+			want: "127.0.127.2",
+		},
+		{
+			name: "UDP address",
+			addr: &net.UDPAddr{
+				IP:   net.ParseIP("127.0.0.1"),
+				Port: 448,
+			},
+			want: "127.0.0.1",
+		},
+		{
+			name: "Other address",
+			addr: &net.UnixAddr{
+				Name: "foobar",
+			},
+			want: "foobar",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := baseIP(tt.addr); got != tt.want {
+				t.Errorf("baseIP() = %v, want %v", got, tt.want)
 			}
 		})
 	}
